JVMCBC-673: Handle error map error codes like core 1.x
Motivation
----------
If we receive an error code that is not known, we can use the
error map and try to derive some information from it that gives
us more information on what's going on underneath.

NOTE: this changeset is NOT handling retry attributes as it does
in 1.x, because how to do it precisely is still under consideration
in the SDK RFC and will be handled at a later point if needed.

Change-Id: Iaab9a84640432f4b07355c6b6bbd6d86178110bb
Reviewed-on: http://review.couchbase.org/112724
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>