JVMCBC-705: Reconnect endpoint if channel closes.
Motivation
----------
This basic functionality has not been implemented until yet: if
the underlying channel closes (usually because the server restarts
or the socket is cut off for some other reason) then the endpoint
needs to immediately reconnect in order to keep going.

Modifications
-------------
The approach we follow here is very similar to how it works in 2.x:
when the channel is inactive the handler calls "out" to the endpoint
signaling this state, which  then in turn triggers the reconnect.

An integration test has been added where a mock netty server simulates
the socket close and we assert that the reconnect happens as well as
that the states are transitioned in the right order.

Change-Id: I9261211f1a5bbbee1149c62946411508e45679ad
Reviewed-on: http://review.couchbase.org/111866
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>