JVMCBC-706: Consolidate KV IO flush when possible.
Motivation
----------
Since the central ringbuffer is gone in core-io 2.x (which also
controlled the central flush), highly concurrent setups we flush
on every request write down into the IO layer. Writes do not trigger
a syscall, but each flush does.

Modifications
-------------
In order to reduce the number of flushes that actually hit the underlying
IO subsystem, netty provides a handler which consolidates flushes when a
read is ongoing (and even if there are just more flushes lined up than
need to be) and flushes once read complete.

Since it is not batching up in the traditional sense (as in it doesn't
actively "wait" for more data) there has been no measurable impact observed
w.r.t increased latency.

Note that an experimental system property has been introduced if we need
to tweak this param for analysis. Setting it to 0 or smaller will even
omit the handler should we need it in the field.

By default it forces a flush after 256 flush attempts, which based on testing
performed quite well in batch scenarios. Smaller batches had slightly less
positive impact and no latency impact was observed with 256 so let's stick
with the default for now.

Change-Id: I32f2d7c53e01712255d20d43185787f806d64f43
Reviewed-on: http://review.couchbase.org/112029
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>