JVMCBC-707: Add async and reactive shutdown methods to env and java apis
Motivation
----------
In some scenarios, it is important that we provide non-blocking access
to the various shutdown methods. Since we already do it async in the
background, it is a matter of exposing those APIs both on the environment
and on the async/reactive clusters. In addition, custom overloads for
timeouts have been added so it can be overridden if needed.

Change-Id: I3a1857af93735c0621e6a656f50d6fea2de901d4
Reviewed-on: http://review.couchbase.org/112128
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>