JVMCBC-707: Defer core shutdown until all nodes have been cleared.
Motivation
----------
Previously, as soon as the config provider would send out the shutdown command
it would return and we would return shutdown complete to the caller of Core
shutdown. This is a problem, because usually afterwards the environment is
cleaned up too on which the core depends. In practice this resulted into
a race condition where the core is still trying to reconfigure/clean up
the nodes, services and endpoints and someone else kills the environment
already.

Modifications
-------------
This changeset:

 - first, fixes a regression introduced with the global config which prevented
   the global config being cleared on shutdown, leading to nodes not being
   cleaned up at all.
 - now in core waits for all nodes to be cleared at a 10ms interval before
   continuing with the shutdown complete mono to the caller.

If all nodes are cleaned up we can be certain that no endpoint or so is trying
to reconnect or use the event loop since they all got the disconnect signal
already.

Change-Id: Ib58ac4de971d209c5c53d8f00c6104f4fa795f83
Reviewed-on: http://review.couchbase.org/112129
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>