JVMCBC-710: Fix concurrent race condition for non-pipelined HTTP endpoints.
Motivation
----------
During concurrency testing it was found that there is a race condition in the
non-pipelined HTTP endpoints where it might be free when checked at the higher
layers but then inside the channel multiple requests might end up there.

Modifications
-------------
Since we do not want to introduce locking in the hot code path, once we are
inside the (single threaded) event handler, check again before writing the
op and if it's occupied retry it immediately, giving it a chance to be sent
to an actual free endpoint.

Integration test has been added, I had to modify some other tests to pass
properly (since the handler now checks the passed in endpoint if it is
pipelined, so there is no free pass in passing in null in the tests
anymore).

I also had to fix the query integration tests since both classes now raced
on creating the primary index on the bucket - this has been refactored
into the parent class.

Change-Id: I20d4b84ec9749fb0b03f05d9a73baa999a9eaab4
Reviewed-on: http://review.couchbase.org/112395
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>