JVMCBC-711: Log unknown KV response codes.
Motivation
----------
It makes sense that if we encounter a response code in the KV service,
we would only propagate UNKNOWN up the stack so let's log it at INFO
level. This gives us a chance to grab it from users logs and see if we
need to add a code or have a deeper problem somewhere going on.

Change-Id: I1c512e1edfee92c41e6889120664367b7e72eb1f
Reviewed-on: http://review.couchbase.org/112460
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>