JVMCBC-703: Short circuit durable write when not available.
Motivation
----------
When a user performs a mutation with durable writes configured, then
we need to short circuit it if it's not available on the bucket in
the first place (i.e. an older server is used).

Modifications
-------------
This changeset introduces an interface that is attached to all mutation
operations which support durable writes. The KV locator checks for their
presence and if it's not in the config as a bucket capability rejects
the request immediately.

Note that the bucket capabilities are now exposed and converted to an
enum set so their lookup is very efficient.

Added a test case to cover this scenario.

Change-Id: I6bbcf0fd7385327ecca0ab159474ed0312b67e19
Reviewed-on: http://review.couchbase.org/112673
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>