JCBC-1375 SDK3: Implement User and Group management API
Motivation
----------
Server responses prior to 6.5 do not include role origins, since there
are no groups and all roles are implicitly innate to the user. This
caused the User objects returned by the user manager to not have any
roles, since only roles with an explicit origin of type="user" would
get transferred from the UserAndMetadata object to the User object.

Modifications
-------------
If a role has no origins, add a synthetic origin with type="user"
so the UserManager behaves the same regardless of Couchbase Server
version.

Change-Id: I7391a76653c8ac99063a225559b4fc7b7aee1eea
Reviewed-on: http://review.couchbase.org/112694
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>