JCBC-1385: Make defaultScope explicit
Motivation
----------
Since only the default scope can have a default collection, it needs
to be accessible on the one hand, but it should also not be possible
to open a default collection from a non-default scope.

Modifications
-------------
This changeset adds the defaultScope to the bucket methods and also removes
the public accessors to the defaultCollection on the scope. It can only be
reached (per RFC) from the bucket level, and checks have been put in place
if someone should try to open with _default manually it rejects that call.

Tests added to verify the rejection on various levels.

Change-Id: I8b302e98975ee5fbb2749f7e7256cee18524fdd5
Reviewed-on: http://review.couchbase.org/112713
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>