JVMCBC-721: Support for view management requests.
Motivation
----------
In preparation to support view management operations in the higher
level bindings, the core needs to support non-streaming requests
against the view endpoint.

Modifications
-------------
This changeset modifies the view endpoint similar to the search
endpoint so that both streaming and non-streaming requests are
possible.

An integration test has been added to test a generic request
against the / endpoint. Note that while the mock has basic view
support it does not support this endpoint so we are ignoring
it there.

Change-Id: I9164f098807ca812df8db7d9e809e5da8994a0fc
Reviewed-on: http://review.couchbase.org/113023
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>