JCBC-1388 SDK3: Implement View management API
Modifications
-------------
Add [Async]ViewIndexManager and related classes.

The following changes were made to support the new view manager:

The view manager needs to know the HTTP status code for failed view
responses so it can throw an appropriate exception. In order to
expose the status code to the View Manager, add HttpStatusCodeException
and set it as the cause of ViewServiceException.

Modify ViewServiceException and SearchServiceException constructors
to accept a cause, and streamline message generation.

Change-Id: Iaa4f424e3858e14359e0ade63b32150aab359a66
Reviewed-on: http://review.couchbase.org/113055
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>