JVMCBC-615: Support for alternate addresses
Motivation
----------
In order to allow the SDK to connect to containerized environments -
like a kubernetes cluster - from the outside, we need to properly
use the alternate address config section that is advertised through
the cluster configuration.

Modifications
-------------
This changeset adds the alternate address piece to the whole config
fetching, updating and reconfiguration step. It works similar to the
implementation in SDK 2, but has been simplified in a couple of files
since SDK 3 was already prepared for this kind of change.

In particular, it adds the alternate identifier to the core context,
which is populated from the config provider and then used from the
core on reconfiguration. In SDK 2 this had been stored in a mutable
fashion on the config itself, but now we can keep the config immutable
and also have the added benefit that the external identifier is logged
as part of the core context logging wherever needed to aid debugging.

More unit tests have been added to make sure the right external identifier
is selected, although at this point I have not been able to add integration
tests since it is hard to get this setup going without an automated
containerized solution that we do not have at this point.

Change-Id: I070af3f258496ee861eddf315e1f5c2e0f0f7fed
Reviewed-on: http://review.couchbase.org/113059
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>