JVMCBC-723: Allow to disable DNS SRV
Motivation
----------
If the given connection string qualifies for DNS SRV lookup, it might still be
the case that a transparent DNS SRV lookup is not a good idea (for example if
the DNS request takes very long). In any case, it should be possible to disable
DNS SRV via a hard switch in the field - but keep it enabled by default.

Change-Id: Ib60d14de8aecc96e8bb21ae8b01589c20d3df9ee
Reviewed-on: http://review.couchbase.org/113091
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>