JVMCBC-600: Expose clientContext for java user options.
Motivation
----------
In order to improve debugability and allow users to attach custom
metadata to requests, this changest exposes the request context
user data to the CommonOptions and (for now) exposes it to the get
operation.

This is the first in a range of changes to improve context-specific
logging and debugging capabilities.

Change-Id: Idcc193448c96db91bc5e27374e9c09fb49635b65
Reviewed-on: http://review.couchbase.org/113125
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>