JVMCBC-600: Support for Mapped Diagnostic Context in Logging
Motivation
----------
SLF4J and some of the underlying implementations have support for
the "mapped diagnostic context", which allows to supply contextual
information that is included in the log output.

In order to use this from the SDKs, we can utilize the clientContext
and if present include it in the request output.

Modifications
-------------
When the logger receives an event and it has a RequestContext attached,
the code checks if the diagnostics are enabled (false by default) and if
so atteches it to the MDC of the logger. The MDC is usually implemented
as a thread local and will clear the previous entry automatically.

Result
------
When the user attaches clientContext it is now automatically present
in the MDC when being logged.

Change-Id: Icd5ab41c31c850cfd149e2f07a8267badd5e1753
Reviewed-on: http://review.couchbase.org/113153
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>