JVMCBC-725: Chunked handler should not close channel when being removed.
Motivation
----------
It looks like this is a leftover line from a refactoring which is now
causing problems together with the chunked/non-chunked switcher.

Modifications
-------------
If the handler is deployed standalone, there is no harm in calling channel
inactive when being removed from the channel, but now with the swticher
in place, the handler being removed is a totally legit action that can
happen at runtime.

Note that with recent changes, the channelInactive is cought by a parent
handler which signals this state to the endpoint anyways, so we are not
loosing any functionality here.

Change-Id: I5b2853f98275fabde96fac6903d845af7ba7bda6
Reviewed-on: http://review.couchbase.org/113170
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>