JCBC-1390: Use only one durability overload.
Motivation
----------
While the SDK was previously following the RFC, it has been decided that
for languages that support overloads it makes sense to just have one
durability() overload in the options. This reduces confusions and makes
it clear that only one can be used at the same time (we still enforce
the check at runtime to make sure).

Change-Id: Ic63d67233dab038567da0ce58a05013740fd714b
Reviewed-on: http://review.couchbase.org/113175
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>