JVMCBC-726: Require a RetryReason when retrying in core.
Motivation
----------
The RetryOrchestrator is the central place where all requests are
passing through, no matter if they are retried or cancelled (based
on the retry strategy configured).

To increase debugability, specifying a reason from the various places
in the stack helps with visibility in the logs why a request has been
retried, and not just when and how often.

Modifications
-------------
This changeset introduces a RetryReason enum and enumerates all the
different (current) possibilities. It then makes it a required argument
on the orchestrator and changes all APIs that call into it to specify
their reason.

Also, the RequestRetriedEvent now has the reason as part of its message,
so with every log line we will also know why it got retried and it allows
us to much quicker pinpoint a potential root cause.

Change-Id: Ia8bf72fd43ac3538f6bc89bda87b08c0ec3e439b
Reviewed-on: http://review.couchbase.org/113177
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>