testing: introduce awaitility and clean up duplicate await calls.
Motivation
----------
Our await method works, but is too simplistic that it will never bail out
should the condition never become true.

Modifications
-------------
This changeset introduces the awaitility library as a test dependency and
uses it (for now) transparently behind our before simplistic method.

In the future it can be extended and/or used directly if needed.

Also note that the same method had been exposed both in test-utils and in
the core, so this changeset removes it in core and adapts the imports
appropriately so there is only one canonical implementation that gets used
across the SDK tests.

Change-Id: I5363a5ed8e4a6bedc742db949a1842ef8cc80220
Reviewed-on: http://review.couchbase.org/113258
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>