JVMCBC-727: Fail cluster-level queries quickly if not serviceable.
Motivation
----------
Cluster-level queries can be performed when one of two criteria
are met:

 - at least one bucket open
 - Couchbase Server 6.5 and later is used (supports global configs)

Modifications
-------------
This changeset performs an explicit check for all operations that can
be performed on the cluster level to see if they can be dispatched
and if not fails fast so the user knows what's going on and doesn't
have to wait until the timeout hits.

One integratin test has been added to verify.

Change-Id: I7ea8162e818f45482c4d7377043802e5a6a2a1a5
Reviewed-on: http://review.couchbase.org/113335
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>