SCBC-35: Change Cluster.connect to return Try[Cluster]
Scala uses functional error handling rather than exceptions.  E.g. if
a method returns a Try[Whatever] it can return a failure, but if it returns
just a Whatever, it *has* to return a Whatever.

Until now the plan was to have opening all resources (cluster, scope etc.)
return a Whatever rather than a Try[Whatever], and if anything went wrong,
continue and return that wrongness on the first real operation tried.
This is because most failures require an initial connection to the
cluster.

However then we hit the case of, what if they've specified a bad
connection string or similar error that can be caught quickly.
So, anything that creates an Environment or Cluster is being changed
to return a Try.

Change-Id: I89d11579467f72ed85e73cedab966f1fbc74e33f
Reviewed-on: http://review.couchbase.org/113337
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>