JVMCBC-729: Implement query fast prepare support
Motivation
----------
Mad Hatter introduces a "fast prepare" support into the query
engine, which allows the SDK to perform a query and execute in
one shot, reducing one network roundtrip for each fresh non-adhoc
query. This is especially important for users who write bad
non-adhoc queries that end up having to prepare all the time.

Modifications
-------------
One change had to be made to the core in that a new "prepared" name
has been added to the potential fields, which is then used to extract
the cache name.

The java-client query accessor has been modified so it - if enabled -
uses the new fast prepare approach and returning the results directly
without going through another execute step in the first place.

No new integration tests have been added since they are covered already
with the "old" prepare support we have for pre 6.5 clusters. They run
automatically when used against a 6.5 cluster and later.

Change-Id: I4d02b4553141be7e24a5e797d55e38e98978974f
Reviewed-on: http://review.couchbase.org/113494
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>