SCBC-59: Add user and groups management API
(Also patches a couple of tiny nits in the Java version.)

As title, adds the user and groups management APIs, based on the RFC and Java
implementation.

I've also made upickle a dependency of the Scala client.  I'm happier using
it over Jackson after a) hitting some (probably solvable) issues with using
Jackson with Scala, as it's not very Scala-native, and b) after benchmarking
many Scala JSON libraries, upickle seems much faster than Jackson, and
though it's not the absolute fastest, it is the one with all the features I
need.

This is a largish changeset already so I've left a) documentation and b)
blocking and async APIs and c) being able to actually access the
ReactiveUserManager from anywhere, for now.  Will get those done after
this review.

Change-Id: If39a985a6e98f351bf001351266eea7ddcfcd821
Reviewed-on: http://review.couchbase.org/113499
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>