Gardening: Rename NetworkResolution.custom(String) to valueOf(String)
Motivation
----------
When reading config values from textual sources (like connector config,
for example), need a nice way to convert a null or empty string
into the default value of AUTO.

Rationale for changing the name from custom() is that you can
also use this method to create any of the standard values by passing
"external", "auto", or "default".

The name valueOf() was selected because it's one of the conventional
names for a method that creates an object from a string.

Modifications
-------------
Rename NetworkResolution.custom to valueOf, and default to AUTO if
the given name is null or empty.

In the NetworkResolution c'tor, require name to be non-null.

Change toString() so it just returns the name, since there's no other
information.

Make the public static constants final to prevent funny business.

Change-Id: Ib27d52a01ba32c408b0756c912cfcecc63f4718d
Reviewed-on: http://review.couchbase.org/113532
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>