core: fix stuck manager message handler test
This changeset fixes a case where the test randomly hangs. Tracked
it down to the situation where the chunk has more than one config
in it, so only ever one would be emitted. Changed the code so it
emits configs as long as they are available in the config chunk
in the handler.

Change-Id: I5d77a2c22468e924e07a17b70e53bd7352ec93d6
Reviewed-on: http://review.couchbase.org/113580
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>