JVMCBC-733: Only dispatch view requests to nodes with primary partitions
Motivation
----------
This changeset ports the same functionality from SDK 2 which is important
to not cause disruption during rebalance. Without this change, during
rebalance the SDK would dispatch requests to a view node without active
KV partitions that leads to a subsequent view engine error.

Modifications
-------------
The config itself had the "node has primary partitions" on it since porting
it from SDK2, but the actual locator dispatch logic was not calling the method
and missing the check.

Now only if a node has a view service enabled AND has primary kv partitions
on it, it can be used for dispatching.

Change-Id: Ic95b3a14f1906ac3dcca4e069a3a6f63cac1fbf2
Reviewed-on: http://review.couchbase.org/113784
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>