JCBC-1391: Fix idempotency for search, view and n1ql queries.
Motivation
----------
The GenericSearchRequest should not always be idempotent, since it depends
on the type of requests.

A view query request (not the management one) is always idempotent.

Also, the N1ql query is idempotent if the readonly flag is set, which by
default is not.

Modifciations
-------------
Request types are adjusted to respect the idempotency flag based on
the property being set, and the higher level code sets it accordingly
when calling their constructors now.

Change-Id: Icf404ddfa7cf244bf8ee1cbf3fb61a3fdbcd4ba6
Reviewed-on: http://review.couchbase.org/114020
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>