SCBC-68: Add query service fast prepare support
Motivation
----------
Mad Hatter introduces a "fast prepare" support into the query
engine, which allows the SDK to perform a query and execute in
one shot, reducing one network roundtrip for each fresh non-adhoc
query. This is especially important for users who write bad
non-adhoc queries that end up having to prepare all the time.

Modifications
-------------
Essentially the Java approach was duplicated.  If the cluster
supports enhanced prepare requests then the prepare is sent
with the auto_execute flag set.  The resulting plan name is
saved and the results returned.

As with Java, no new tests are added as the existing tests will
cover this.

Change-Id: I5e7bf2c9f5d7cf8400d620f6cbf8e25a1de84670
Reviewed-on: http://review.couchbase.org/114055
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>