JVMCBC-735: Close KV connection on certain response status codes.
Motivation
----------
Looking into the KV error map, there are a couple of response status
codes which indicate that a connection need to be reset immediately
in order to have a chance to get back to a sane state.

Modifications
-------------
This changeset adds the functionality to close the KV connection
proactively on EINTERNAL, NO_BUCKET and NOT_INITIALIZED, in-line
with the KV error map recommendations.

Note that the surrounding code has been refactored and extended
for clarity and additional debugging (i.e. the closing of the
socket has been refactored into its own method and is now called
from a couple places with explicit reasons).

Change-Id: Ic8401665e8dad0e5d262ac9aba683c261dae121c
Reviewed-on: http://review.couchbase.org/114111
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>