JVMCBC-736: Record and dump retry reasons
Motivation
----------
When a retry happens, the code should not just record that an attempt happened,
but also store a set of retry reasons so that when debugging it is possible
to figure out why the retry happened in the first place.

Modifications
-------------
The RequestContext now demands providing a reason when incrementing the retry
attempts and lazily populates an EnumSet with reasons. The orchestrator now
passes the reason down to the context.

Change-Id: Ie79e4b2aeb61ef8889f223a7f32dc0a0dc56bbeb
Reviewed-on: http://review.couchbase.org/114113
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>