JVMCBC-1391: Transparently retry certain KV error codes
Motivation
----------
As defined in the retry handling RFC, certain KV error codes should
be sent into retry rather than completing the op directly. This
changeset modifies the error handling in the KV layer so that a
check for those status codes is added and then it is sent into the
retry process with the corresponding reason.

Change-Id: Ib243db2f2f0a52abb32e36613c26852fcdd38748
Reviewed-on: http://review.couchbase.org/114115
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>