JVMCBC-737: Harden HTTP handlers for write race conditions.
Motivation
----------
During manual code inspection, certain shortcomings of the http-based
handlers have been found which are addressed as part of this changeset.

Modifications
-------------
The following changes have been made in this changeset:

 - When a write is retried because of a concurrent request condition,
   the outstanding requests counter needs to be decrement so it can
   eventually come back down to 0.
 - Write failures are now uniformly handled in try/catch blocks and also
   decrement the outstanding request counter on the endpoint.
 - A utility method/class has been introduced to close the socket proactively
   with a given reason since it is needed in more than one handler now.

Tests had to be adapted/added to reflect the new situation and also to make
sure the "write race" really decrements the outstanding request counter.

Change-Id: Ifc2c2fec7759400cbc807e2efe5bb8fe5b015ef3
Reviewed-on: http://review.couchbase.org/114125
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>