JCBC-1438: Refactor Encoding and Decoding
Motivation
----------
The encoding and decoding story of JSON and non-JSON content
has been evolving for a while but the SDK hasn't followed
suite w.r.t changes to the SDK spec. This changeset aims to
remedy this, although there are follow-up changes expected
that carry those changes into other APIs as well.

Modifications
-------------
The big theme of this changeset is to get rid of the generic
encoders and decoders and replace them with:

 - Serializers, handling the raw byte transformations (i.e.
   JSON and java object serialization)
 - Transcoders, which utilize Serializers for the heavy lifting
   and then marry it up with the Common flags / DataFormat under
   the KV umbrella.

The DataFormat now acts as an abstraction over the common flags
and allows the user to specify it on the KV options. This has the
benefit that there is no need anymore for content wrapper classes
and runtime inspection, but rather the user telling the SDK explcitly
what type of data format they are passing in (and then the SDK selecting
the right serializer for it).

Tests had to be reworked, deleted and added in order to fit the
new structure.

Change-Id: If3015c42f0a1267f6766c1ea4f9f4a8a85f7f949
Reviewed-on: http://review.couchbase.org/114130
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>