JCBC-1432 LazyCircuitBreaker does not handle System.nanoTime overflow
Motivation
----------
Incorrect System.nanoTime comparisons could lead to unexpected results
as the value approaches Long.MAX_VALUE.

Modifications
-------------
Initialize the timestamps so they are always in the distant past
regardless of System.nanoTime's reference point.

Compare nanoTime values by measuring the result of subtracting one from
the other.

Avoid calling System.nanoTime if the result is not used by a code path
(it's not super-expensive, but it's not free).

Change-Id: Ie9f508678c9b590669eb67e7d2ca724ecdc01ffc
Reviewed-on: http://review.couchbase.org/114191
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>