JVMCBC-739: Ignore NOT_FOUND global config load responses.
Motivation
----------
In the situation where the server supports global configs but does
not have one assigned from ns_server set, kv_engine reports back
with not found (0x01). During loading there is not much we can do
about it, eventually the global config poller will pick them up
once they arrive.

Modifications
-------------
The global loader now raises the not found event explicitly, and
the core handles it and dumps a specific reason in the log to
identify this situation.

Also the code is changed so that no failure during the global loading
actually holds up the bootstrapping, since "gcccp" is optional and
the situation might resolv itself. Any unexpected errors are in found
in the logs at WARN level.

Change-Id: I54936175f2305c966bb4e07b18d17078d5828709
Reviewed-on: http://review.couchbase.org/114530
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>