JCBC-1441 Expose datastructures on the blocking collection API
Motivation
==========
Allow users to create new (or use existing) datacollections directly
from the connection object, and specify an Options object for things
like the number of cas retries, timeout, etc...

To construct one of these, I chose a constructore that expects a required
Class<T> entityTypeClass, then an optional Options which is specific to
the datastructure class.

There are other possibilities, so lets decide which way to go, and I can finish this off.

Modification
============

Change-Id: I0ef36ed8451a18b3bb537475b362bd5d25a01852
Reviewed-on: http://review.couchbase.org/114545
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>