core: add flag to enable experimental unordered execution.
Motivation
----------
This changeset adds an undocumented flag to enable the (still
experimental) hello flag for unordered execution. To detect this
at runtime, a WARN level event is raised.

Note that at the moment the serer does NOT support this properly,
so only enable the flag if you know exactly what you are doing!

THERE WILL BE DRAGONS!

Change-Id: Idded09f559379d13c5b7ac282f82306994e6c4e4
Reviewed-on: http://review.couchbase.org/114621
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>