JVMCBC-741: Check for low timeouts before writing onto the network.
Motivation
----------
If users set timeouts below our wheel timer tick interval (100ms),
even if they shouldn't, we need a way to double check that the op
isn't timed out already and if so stop processing it.

Modifications
-------------
We do this by storing the absolute timeout deadline when the request
is created and check every time before we do the IO write (which is
the most resource intensive part of the dispatch stack).

Added a test to make sure this works as expected.

Change-Id: I6bc65941cf25148ae7ac3777d3cbde247014511d
Reviewed-on: http://review.couchbase.org/114702
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>