JCBC-1441 Expose datastructures on the blocking collection API
Motivation
==========
Add api calls to allow datastructures to be created from the collection
object.

Modification
============
This is a followup to address a couple issues with datastructures.  The
main thing is to be sure to pass all of the CommonOptions through when
making calls.  To do so we now make a copy of the option coming in, and
create new options specific to the operation we are performing when needed.

While there, fixed a bug with peek() in an empty queue, found a few places
we needed options, etc...

Change-Id: Ia2ad40f8eabfa2482c3318a5a6adf383438457f3
Reviewed-on: http://review.couchbase.org/114707
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>