JVMCBC-742: Rework Credentials to proper Authenticator
This changeset brings the credentials in line with the RFC which calls
it authenticator. It also adds the CertificateAuthenticator in preparation
to support client certificate authentication in a later commit.

Note that previously this code acted as a pure credential store, but now the
logic is inverted so that the authenticator actually performs certain auth
operations (i.e. adding sasl auth to the kv pipeline or basic auth creds to
the http requests). This leaves the door open for future authenticator
implementations which could even do things dynamically if needed.

Also, the PasswordAuthenticator is extended with a Supplier for both username
and password so that systems which fetch it dynamically (even at runtime) can do
so.

Change-Id: I830600031915b17a060f21b3b2ae1ed55ee706ee
Reviewed-on: http://review.couchbase.org/114719
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>