java: expose core and client hash and version on the environment.
Motivation
----------
In environments like QE it is important to have access to both the
client and core version & hash to programmatically return it to the
calling system so that it's clear what exact version is used.

The code already had this information private, all we need to do
is expose it as public methods.

Change-Id: Iba4b5f38eb029bb658ed622f714c00bb2154a6f3
Reviewed-on: http://review.couchbase.org/114741
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>