JCBC-1448: Allow to customize global transcoder and serializer
Motivation
----------
Previously the default transcoder and serializer used to be static
instances that couldn't be customized. Based on feedback it makes
sense that both are configurable on the environment and then act
as the defaults across the client.

Modifications
-------------
This changeset introduces default transcoder and serializer as part
of the Cluster environment and then removes the static INSTANCE variables
of them.

The rest of the code is pure refactoring to bring the defaults in scope
where needed.

Change-Id: Id774d509dfd2bae656e28002a9b996ed4dd8b42d
Reviewed-on: http://review.couchbase.org/114745
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>