java:  Rework Search API
Motivation
----------
This changeset heavily reworks the Search API and brings
it closer to what is going to be in the Search RFC
eventually. Expect more changes down the road but this
should be pretty close hopefully.

Note: still pretty rough in the options and the result, but
the overall structure of the query should be alright.

Change-Id: I884805e72ddcb2ecb5c2da1bf2a3049c46d5ef18
Reviewed-on: http://review.couchbase.org/114803
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>