JCBC-1451 Review for log redaction compliance
Motivation
==========
We want to insure we have redacted throughout the sdk3 code, as there
was quite a bit of new code added.  So lets sweep through the code and
look for likely places, adding/updating as needed.

Modification
============
Added redaction in several places.  Also, changed the redaction in a few
to follow (my interpretation of) the spec.

Change-Id: Ie7a2597f0c1a4cd3b4888f92ae447d0e8452d8c1
Reviewed-on: http://review.couchbase.org/114911
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>