JCBC-1452: Expose ClusterOptions on connect
Motivation
----------
This is the first in a series of chaanges to decouple the seed nodes
and the authenticator from the environment and instead have it as
part of options to connect (including the environment).

This brings the SDK in-line with the RFC, but also uses the consistent
pattern of "options", makes it possible to share environments across
clusters if needed and also opens the door for future options to be
applied on cluster connect.

Modifications
-------------
This first change:

 - Adds the ClusterOptions class
 - Introduces it as an overload for Cluster and its async/reactive
   variants
 - Adapts the test suite to use the new connect methods

Note that right now authenticator and seed nodes are set on both
the environment and the options, but in future commits they will
go away on the environment.

Change-Id: I75d1a1602fd88dccdbe09b312375372e58d20a37
Reviewed-on: http://review.couchbase.org/114952
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>